Argument passing, context switching, cleaned up main
A simple burn-in loop was placed in first(). activate() was moved into the
assembly file. Kernel and user can now pass a single argument back and forth.

The user context stack bug was in swtch() -- erroneously adding to the stack
without decrementing it. What is still needed is saving of the Program Status
Register for user/kernel space, also perhaps some optimizing of the context
switch instructions. GO-GO usermode functions implementing!