New code: shell of system calls
User code has been moved to user.c. The user's "system calls" are now in
usyscall.{s,h} (usyscall.s is ugly, there is maybe a better way to do this?).
The kernel's helper functions that perform the syscall services are named with
a _ before them, so that we know which is which: see ksyscall.h. The idea is
that the kernel will look at the argument to SWI, then call the appropriate
helper function. I will soon add support for this to the exception handler
dealy.

I've also added some definitions in switch.h for kernel structure stuff, this
will most definitely need to be revised.